Update v3.0 build 3
* add some config keys
* add defualt classes
* change world handling
* remove guard delpath command (replaced)

+++ myrpg/src/org/caliog/myRPG/Group/Group.java
@@ -38,7 +38,7 @@ public class Group {
 	if (this.players.containsKey(member)) {
 	    return false;
 	}
-	this.players.put(member, Float.valueOf(0.0F));
+	this.players.put(member, 0.0F);
 	updateScoreboard();
 	return true;
     }
